fix: PR id vs number confusion and integer overflow

+++ src/schema/tables/pull-request.table.ts
@@ -2,7 +2,7 @@ import { Column, PrimaryColumn, Table } from '@immich/sql-tools';
 
 @Table('pull_request')
 export class PullRequestTable {
-  @PrimaryColumn({ type: 'integer' })
+  @PrimaryColumn({ type: 'bigint' })
   id!: number;
 
   @Column()
